split client tests
