+++ package/src/client.spec.ts
@@ -5,12 +5,10 @@ import "@testing-library/jest-dom/extend-expect";
 import "@testing-library/jest-dom";
 import "jest-expect-message";
 import { renderHook } from "@testing-library/react-hooks";
-import { useFlags, cache, UseFlagsOptions } from "./client";
-import { configure, _resetConfig } from "./config";
+import { useFlags, cache } from "./client";
+import { _resetConfig } from "./config";
 import * as fetchMock from "fetch-mock-jest";
 import { deleteAllCookies } from "../jest/delete-all-cookies";
-import { nanoid } from "nanoid";
-import { FlagBag, Flags, InitialFlagState } from "./types";
 
 beforeEach(() => {
   _resetConfig();
@@ -28,803 +26,4 @@ it("it warns on missing config", () => {
   expect(result.error).toEqual(
     Error("@happykit/flags: Missing configuration. Call configure() first.")
   );
-});
-
-describe("client-side rendering", () => {
-  describe("when cookie is not set", () => {
-    it("generates a visitor key", async () => {
-      let visitorKey;
-      // mock incoming post request...
-      fetchMock.post(
-        // ...when it matches this...
-        {
-          url: "https://happykit.dev/api/flags/flags_pub_000000",
-          body: {
-            // there is a visitorKey but we don't know the value as it's generated
-            // visitorKey: expect.any(String),
-            user: null,
-            traits: null,
-            static: false,
-          },
-          matchPartialBody: true,
-        },
-        // ... and respond with that...
-        (url: string, options: RequestInit, request: Request) => {
-          // parse visitorKey so we can mirror it back
-          const body = JSON.parse(options.body as string);
-          visitorKey = body.visitorKey;
-          expect(typeof visitorKey).toBe("string");
-
-          return {
-            flags: {
-              ads: true,
-              checkout: "medium",
-              discount: 5,
-              purchaseButtonLabel: "Purchase",
-            },
-            visitor: { key: visitorKey },
-          };
-        }
-      );
-
-      configure({ envKey: "flags_pub_000000" });
-      expect(document.cookie).toEqual("");
-
-      const { result, waitForNextUpdate } = renderHook(() => useFlags());
-
-      expect(result.all).toHaveLength(2);
-
-      await waitForNextUpdate();
-
-      expect(result.all).toEqual([
-        {
-          flags: null,
-          data: null,
-          error: null,
-          fetching: false,
-          settled: false,
-          visitorKey: null,
-          revalidate: expect.any(Function),
-        },
-        {
-          flags: null,
-          data: null,
-          error: null,
-          fetching: true,
-          settled: false,
-          visitorKey: visitorKey,
-          revalidate: expect.any(Function),
-        },
-        {
-          flags: {
-            ads: true,
-            checkout: "medium",
-            discount: 5,
-            purchaseButtonLabel: "Purchase",
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: "medium",
-              discount: 5,
-              purchaseButtonLabel: "Purchase",
-            },
-            visitor: { key: visitorKey },
-          },
-          error: null,
-          fetching: false,
-          settled: true,
-          visitorKey: visitorKey,
-          revalidate: expect.any(Function),
-        },
-      ]);
-
-      expect(
-        (result.all[1] as FlagBag<any>).visitorKey,
-        "visitor key may not change"
-      ).toEqual((result.all[2] as FlagBag<any>).visitorKey);
-    });
-  });
-
-  describe("when cookie is set", () => {
-    it("reuses the visitor key", async () => {
-      // prepare cookie before test
-      const visitorKeyInCookie = nanoid();
-      document.cookie = `hkvk=${visitorKeyInCookie}`;
-
-      // prepare response before test
-      fetchMock.post(
-        {
-          url: "https://happykit.dev/api/flags/flags_pub_000000",
-          body: {
-            visitorKey: visitorKeyInCookie,
-            user: null,
-            traits: null,
-            static: false,
-          },
-          matchPartialBody: true,
-        },
-        {
-          headers: { "content-type": "application/json" },
-          body: {
-            flags: {
-              ads: true,
-              checkout: "medium",
-              discount: 5,
-              purchaseButtonLabel: "Purchase",
-            },
-            visitor: { key: visitorKeyInCookie },
-          },
-        }
-      );
-
-      configure({ envKey: "flags_pub_000000" });
-      expect(document.cookie).toEqual(`hkvk=${visitorKeyInCookie}`);
-
-      // start actual testing
-      const { result, waitForNextUpdate } = renderHook(() => useFlags());
-
-      expect(result.all).toHaveLength(2);
-
-      await waitForNextUpdate();
-
-      expect(result.all).toEqual([
-        {
-          flags: null,
-          data: null,
-          error: null,
-          fetching: false,
-          settled: false,
-          visitorKey: null,
-          revalidate: expect.any(Function),
-        },
-        {
-          flags: null,
-          data: null,
-          error: null,
-          fetching: true,
-          settled: false,
-          visitorKey: visitorKeyInCookie,
-          revalidate: expect.any(Function),
-        },
-        {
-          fetching: false,
-          flags: {
-            ads: true,
-            checkout: "medium",
-            discount: 5,
-            purchaseButtonLabel: "Purchase",
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: "medium",
-              discount: 5,
-              purchaseButtonLabel: "Purchase",
-            },
-            visitor: { key: visitorKeyInCookie },
-          },
-          error: null,
-          settled: true,
-          visitorKey: visitorKeyInCookie,
-          revalidate: expect.any(Function),
-        },
-      ]);
-    });
-  });
-});
-
-describe("server-side rendering (hybrid)", () => {
-  describe("when visitorKey is not set in cookie", () => {
-    it("uses the visitor key generated on the server", async () => {
-      expect(document.cookie).toBe("");
-      const generatedVisitorKey = nanoid();
-
-      configure({ envKey: "flags_pub_000000" });
-      expect(document.cookie).toEqual("");
-
-      const initialStateFromProps: InitialFlagState<Flags> = {
-        input: {
-          endpoint: "https://happykit.dev/api/flags",
-          envKey: "flags_pub_000000",
-          requestBody: {
-            visitorKey: generatedVisitorKey,
-            user: null,
-            traits: null,
-            static: false,
-          },
-        },
-        outcome: {
-          data: {
-            flags: {
-              ads: true,
-              checkout: "short",
-              discount: 5,
-              purchaseButtonLabel: "Buy now",
-            },
-            visitor: { key: generatedVisitorKey },
-          },
-        },
-      };
-
-      const { result } = renderHook(() =>
-        useFlags({ initialState: initialStateFromProps })
-      );
-
-      expect(result.all).toEqual([
-        {
-          flags: {
-            ads: true,
-            checkout: "short",
-            discount: 5,
-            purchaseButtonLabel: "Buy now",
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: "short",
-              discount: 5,
-              purchaseButtonLabel: "Buy now",
-            },
-            visitor: { key: generatedVisitorKey },
-          },
-          error: null,
-          fetching: false,
-          settled: true,
-          visitorKey: generatedVisitorKey,
-          revalidate: expect.any(Function),
-        },
-      ]);
-
-      expect(fetchMock.calls(), "should not fetch at all").toHaveLength(0);
-
-      // getFlags() would set the cookies in the response to the page itself,
-      // but that server function is mocked in this example.
-      //
-      // The header sent in response to the page load would look like
-      // Set-Cookie: hkvk=generatedVisitorKey; Path=/; Max-Age=15552000; SameSite=Lax
-      //
-      // We expect an empty cookie instead, since this is a unit, in which
-      // the getFlags() on the server does not run, and thus has no chance to
-      // set the cookie.
-      expect(document.cookie).toBe("");
-    });
-  });
-
-  describe("when visitorKey is set in cookie", () => {
-    it("reuses the visitor key", async () => {
-      // prepare cookie before test
-      const visitorKeyInCookie = nanoid();
-      document.cookie = `hkvk=${visitorKeyInCookie}`;
-
-      configure({ envKey: "flags_pub_000000" });
-      expect(document.cookie).toEqual(`hkvk=${visitorKeyInCookie}`);
-
-      const initialStateFromProps: InitialFlagState<Flags> = {
-        input: {
-          endpoint: "https://happykit.dev/api/flags",
-          envKey: "flags_pub_000000",
-          requestBody: {
-            visitorKey: visitorKeyInCookie,
-            user: null,
-            traits: null,
-            static: false,
-          },
-        },
-        outcome: {
-          data: {
-            flags: {
-              ads: true,
-              checkout: "short",
-              discount: 5,
-              purchaseButtonLabel: "Buy now",
-            },
-            visitor: { key: visitorKeyInCookie },
-          },
-        },
-      };
-
-      // start actual testing
-      const { result } = renderHook(() =>
-        useFlags({ initialState: initialStateFromProps })
-      );
-
-      expect(fetchMock.calls(), "should not fetch at all").toHaveLength(0);
-
-      expect(result.all).toEqual([
-        {
-          fetching: false,
-          flags: {
-            ads: true,
-            checkout: "short",
-            discount: 5,
-            purchaseButtonLabel: "Buy now",
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: "short",
-              discount: 5,
-              purchaseButtonLabel: "Buy now",
-            },
-            visitor: { key: visitorKeyInCookie },
-          },
-          error: null,
-          settled: true,
-          visitorKey: visitorKeyInCookie,
-          revalidate: expect.any(Function),
-        },
-      ]);
-    });
-  });
-});
-
-describe("static site generation (hybrid)", () => {
-  describe("when cookie is not set", () => {
-    it("generates a visitorKey", async () => {
-      let generatedVisitorKey;
-
-      fetchMock.post(
-        {
-          url: "https://happykit.dev/api/flags/flags_pub_000000",
-          body: {
-            // false because this is the request of the client afer hydration,
-            // not the one during static site generation
-            static: false,
-            traits: null,
-            user: null,
-            // visitorKey: generatedVisitorKey,
-          },
-          matchPartialBody: true,
-        },
-        (url: string, options: RequestInit, request: Request) => {
-          // parse visitorKey so we can mirror it back
-          const body = JSON.parse(options.body as string);
-          generatedVisitorKey = body.visitorKey;
-          expect(typeof generatedVisitorKey).toBe("string");
-          expect(generatedVisitorKey).toHaveLength(21);
-
-          return {
-            flags: {
-              ads: true,
-              checkout: "short",
-              discount: 5,
-              purchaseButtonLabel: "Purchase",
-            },
-            visitor: { key: generatedVisitorKey },
-          };
-        }
-      );
-
-      configure({ envKey: "flags_pub_000000" });
-      expect(document.cookie).toEqual("");
-
-      const initialStateFromProps: InitialFlagState<Flags> = {
-        input: {
-          endpoint: "https://happykit.dev/api/flags",
-          envKey: "flags_pub_000000",
-          requestBody: {
-            visitorKey: null,
-            user: null,
-            traits: null,
-            static: true,
-          },
-        },
-        outcome: {
-          data: {
-            flags: {
-              ads: true,
-              checkout: null,
-              discount: 5,
-              purchaseButtonLabel: null,
-            },
-            visitor: null,
-          },
-        },
-      };
-
-      const { result, waitForNextUpdate } = renderHook(() =>
-        useFlags({ initialState: initialStateFromProps })
-      );
-
-      expect(result.all).toHaveLength(2);
-      expect(document.cookie).toEqual("");
-      await waitForNextUpdate();
-      expect(document.cookie).toEqual(`hkvk=${generatedVisitorKey}`);
-
-      // the worker never sets cookies now, so the client has to deal
-      // with generating a visitor key and storing it as a cookie
-      //
-      // we need this to work nicely with ssr and ssg, so the expected behaviour
-      // of useFlags depends on which initial state is passed in
-      //
-      // rendered with ssr? => server should have generated visitorKey if not present, and it should be set in response
-      // rendered with ssg? => server sets visitorKey: null in request, and useFlags needs to reevaluate the flags after the initial render
-
-      expect(result.all).toEqual([
-        {
-          flags: {
-            ads: true,
-            checkout: null,
-            discount: 5,
-            purchaseButtonLabel: null,
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: null,
-              discount: 5,
-              purchaseButtonLabel: null,
-            },
-            visitor: null,
-          },
-          error: null,
-          fetching: false,
-          settled: false,
-          visitorKey: null,
-          revalidate: expect.any(Function),
-        },
-        {
-          flags: {
-            ads: true,
-            checkout: null,
-            discount: 5,
-            purchaseButtonLabel: null,
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: null,
-              discount: 5,
-              purchaseButtonLabel: null,
-            },
-            visitor: null,
-          },
-          error: null,
-          fetching: true,
-          settled: false,
-          visitorKey: null,
-          revalidate: expect.any(Function),
-        },
-        {
-          flags: {
-            ads: true,
-            checkout: "short",
-            discount: 5,
-            purchaseButtonLabel: "Purchase",
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: "short",
-              discount: 5,
-              purchaseButtonLabel: "Purchase",
-            },
-            visitor: { key: generatedVisitorKey },
-          },
-          error: null,
-          fetching: false,
-          settled: true,
-          visitorKey: generatedVisitorKey,
-          revalidate: expect.any(Function),
-        },
-      ]);
-
-      // ensure there are no further updates
-      await expect(waitForNextUpdate({ timeout: 500 })).rejects.toThrow(
-        "Timed out"
-      );
-    });
-  });
-
-  describe("when cookie is set", () => {
-    it("generates a visitorKey", async () => {
-      // prepare cookie before test
-      const visitorKeyInCookie = nanoid();
-      document.cookie = `hkvk=${visitorKeyInCookie}`;
-
-      fetchMock.post(
-        {
-          url: "https://happykit.dev/api/flags/flags_pub_000000",
-          body: {
-            // false because this is the request of the client afer hydration,
-            // not the one during static site generation
-            static: false,
-            traits: null,
-            user: null,
-            visitorKey: visitorKeyInCookie,
-          },
-        },
-        {
-          flags: {
-            ads: true,
-            checkout: "short",
-            discount: 5,
-            purchaseButtonLabel: "Purchase",
-          },
-          visitor: { key: visitorKeyInCookie },
-        }
-      );
-
-      configure({ envKey: "flags_pub_000000" });
-
-      const initialStateFromProps: InitialFlagState<Flags> = {
-        input: {
-          endpoint: "https://happykit.dev/api/flags",
-          envKey: "flags_pub_000000",
-          requestBody: {
-            visitorKey: null,
-            user: null,
-            traits: null,
-            static: true,
-          },
-        },
-        outcome: {
-          data: {
-            flags: {
-              ads: true,
-              checkout: null,
-              discount: 5,
-              purchaseButtonLabel: null,
-            },
-            visitor: null,
-          },
-        },
-      };
-
-      const { result, waitForNextUpdate } = renderHook(() =>
-        useFlags({ initialState: initialStateFromProps })
-      );
-
-      expect(result.all).toHaveLength(2);
-      expect(document.cookie).toEqual(`hkvk=${visitorKeyInCookie}`);
-      await waitForNextUpdate();
-      expect(document.cookie).toEqual(`hkvk=${visitorKeyInCookie}`);
-
-      expect(result.all).toEqual([
-        {
-          flags: {
-            ads: true,
-            checkout: null,
-            discount: 5,
-            purchaseButtonLabel: null,
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: null,
-              discount: 5,
-              purchaseButtonLabel: null,
-            },
-            visitor: null,
-          },
-          error: null,
-          fetching: false,
-          settled: false,
-          visitorKey: null,
-          revalidate: expect.any(Function),
-        },
-        {
-          flags: {
-            ads: true,
-            checkout: null,
-            discount: 5,
-            purchaseButtonLabel: null,
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: null,
-              discount: 5,
-              purchaseButtonLabel: null,
-            },
-            visitor: null,
-          },
-          error: null,
-          fetching: true,
-          settled: false,
-          visitorKey: null,
-          revalidate: expect.any(Function),
-        },
-        {
-          flags: {
-            ads: true,
-            checkout: "short",
-            discount: 5,
-            purchaseButtonLabel: "Purchase",
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: "short",
-              discount: 5,
-              purchaseButtonLabel: "Purchase",
-            },
-            visitor: { key: visitorKeyInCookie },
-          },
-          error: null,
-          fetching: false,
-          settled: true,
-          visitorKey: visitorKeyInCookie,
-          revalidate: expect.any(Function),
-        },
-      ]);
-
-      // ensure there are no further updates
-      await expect(waitForNextUpdate({ timeout: 500 })).rejects.toThrow(
-        "Timed out"
-      );
-    });
-  });
-});
-
-describe("stories", () => {
-  describe("client-side rendering", () => {
-    it.only("works", async () => {
-      let generatedVisitorKey = null;
-      fetchMock.post(
-        {
-          url: "https://happykit.dev/api/flags/flags_pub_000000",
-          body: {
-            // there is a visitorKey but we don't know the value as it's generated
-            // visitorKey: expect.any(String),
-            user: null,
-            traits: null,
-            static: false,
-          },
-          matchPartialBody: true,
-        },
-        (url: string, options: RequestInit, request: Request) => {
-          // parse visitorKey so we can mirror it back
-          const body = JSON.parse(options.body as string);
-          generatedVisitorKey = body.visitorKey;
-          expect(typeof generatedVisitorKey).toBe("string");
-
-          return {
-            flags: {
-              ads: true,
-              checkout: "medium",
-              discount: 5,
-              purchaseButtonLabel: "Purchase",
-            },
-            visitor: { key: generatedVisitorKey },
-          };
-        }
-      );
-
-      configure({ envKey: "flags_pub_000000" });
-      expect(document.cookie).toEqual("");
-
-      const { result, waitForNextUpdate, rerender } = renderHook<
-        UseFlagsOptions,
-        FlagBag<Flags>
-      >((options) => useFlags(options), { initialProps: undefined });
-
-      const rendersBeforeFirstSettlement = [
-        {
-          flags: null,
-          data: null,
-          error: null,
-          fetching: false,
-          settled: false,
-          visitorKey: null,
-          revalidate: expect.any(Function),
-        },
-        {
-          flags: null,
-          data: null,
-          error: null,
-          fetching: true,
-          settled: false,
-          visitorKey: generatedVisitorKey,
-          revalidate: expect.any(Function),
-        },
-      ];
-      expect(result.all).toEqual(rendersBeforeFirstSettlement);
-
-      await waitForNextUpdate();
-
-      const rendersOnFirstSettlement = [
-        {
-          flags: {
-            ads: true,
-            checkout: "medium",
-            discount: 5,
-            purchaseButtonLabel: "Purchase",
-          },
-          data: {
-            flags: {
-              ads: true,
-              checkout: "medium",
-              discount: 5,
-              purchaseButtonLabel: "Purchase",
-            },
-            visitor: { key: generatedVisitorKey },
-          },
-          error: null,
-          fetching: false,
-          settled: true,
-          visitorKey: expect.any(String),
-          revalidate: expect.any(Function),
-        },
-      ];
-
-      expect(result.all).toEqual([
-        ...rendersBeforeFirstSettlement,
-        ...rendersOnFirstSettlement,
-      ]);
-
-      fetchMock.post(
-        {
-          url: "https://happykit.dev/api/flags/flags_pub_000000",
-          body: {
-            visitorKey: generatedVisitorKey,
-            user: { key: "george" },
-            traits: null,
-            static: false,
-          },
-        },
-        {
-          flags: {
-            ads: true,
-            checkout: "medium",
-            discount: 10,
-            purchaseButtonLabel: "Purchase",
-          },
-          visitor: { key: generatedVisitorKey },
-        },
-        { overwriteRoutes: true }
-      );
-
-      rerender({ user: { key: "george" } });
-
-      expect(result.all).toHaveLength(5);
-
-      await waitForNextUpdate();
-
-      expect(result.all).toEqual([
-        ...rendersBeforeFirstSettlement,
-        ...rendersOnFirstSettlement,
-        // for some reason this is rendered again, I suspect it's because
-        // of the "rerender" call?
-        ...rendersOnFirstSettlement,
-        {
-          data: null,
-          error: null,
-          fetching: true,
-          flags: null,
-          revalidate: expect.any(Function),
-          settled: false,
-          visitorKey: generatedVisitorKey,
-        },
-        {
-          data: {
-            flags: {
-              ads: true,
-              checkout: "medium",
-              discount: 10,
-              purchaseButtonLabel: "Purchase",
-            },
-            visitor: { key: generatedVisitorKey },
-          },
-          error: null,
-          fetching: false,
-          flags: {
-            ads: true,
-            checkout: "medium",
-            discount: 10,
-            purchaseButtonLabel: "Purchase",
-          },
-          revalidate: expect.any(Function),
-          settled: true,
-          visitorKey: generatedVisitorKey,
-        },
-      ]);
-
-      expect(
-        (result.all[1] as FlagBag<any>).visitorKey,
-        "visitor key may not change"
-      ).toEqual((result.all[2] as FlagBag<any>).visitorKey);
-    });
-  });
 });
